requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
langchain==0.2.3
langchain-community==0.2.3
langchain-groq==0.1.4
huggingface-hub==0.23.1
sentence-transformers==2.6.1
faiss-cpu==1.8.0
docx2txt==0.8
python-docx==1.1.0
fpdf==1.7.2
transformers==4.41.2
PyPDF2==3.0.1
werkzeug==3.0.1
